Add soft shadows
This is hardcoded behaviour which needs to be properly fixes but for now it produces pleasing
results. How it works:

When lambertian material is hit it shoots some amount of rays to a randomized direction
towards each light source and upon receiving them produces hit/total ratio
which is then multiplied by the collision dot product resulting in soft shadows and darker areas
where two shadows "collide".